requirements.txt
fastAPI~=0.95.1
pytest~=7.3.1
uvicorn~=0.22.0
cassandra-driver~=3.27.0
python-dotenv~=1.0.0
pytest-cov
pylint~=2.17.4
black~=23.3.0
pika~=1.3.2
pytest-asyncio
pytest-integration
pytest-envvars
httpx~=0.24.1
asyncio~=3.4.3
argon2-cffi
python-jose[cryptography]
logging-test-case
opencv-python
mtcnn-opencv
starlette~=0.26.1
pydantic~=1.10.7
h11~=0.14.0
pip~=23.3.1
wheel~=0.37.1
cryptography~=40.0.2
rsa~=4.9
pyasn1~=0.5.0
setuptools~=60.2.0
dill~=0.3.6
numpy~=1.26.1
ecdsa~=0.18.0
anyio~=3.6.2
sniffio~=1.3.0
platformdirs~=3.5.1
tomli~=2.0.1
packaging~=23.1
pathspec~=0.11.1
click~=8.1.3
six~=1.16.0
httpcore~=0.17.1
idna~=3.4
certifi~=2023.5.7
isort~=5.12.0
geomet~=0.2.1.post1
astroid~=2.15.5
tomlkit~=0.11.8
mccabe~=0.7.0
pluggy~=1.0.0
exceptiongroup~=1.1.1
iniconfig~=2.0.0
coverage~=7.2.5
pycparser~=2.21
cffi~=1.15.1
